perf: Lazily initialize git2 fetch transports
Depending on how cargo is built and user settings,
this call is slow and noticeable when doing short-lived cargo
invocations.
However, with the sparse registry, most calls are not using git fetch.
If we also migrate to git-cli by default, that will make this never
used.
Let's stop making everyone pay this cost.

A bit shady what we are doing with `unsafe` but it is hard to maintain
this guarentee and lazy initialize.
We could make things more composable for cargo-the-library users by
providing our own lock-wrapper transport registration functions but it
isn't clear we'd get much value out of that work.
