+++ src/cargo/lib.rs
@@ -7,10 +7,16 @@
 //! - <https://doc.rust-lang.org/nightly/nightly-rustc/cargo>: targeted at cargo contributors
 //!   - Updated on each update of the `cargo` submodule in `rust-lang/rust`
 //!
-//! > This library is maintained by the Cargo team, primarily for use by Cargo
-//! > and not intended for external use (except as a transitive dependency). This
-//! > crate may make major changes to its APIs. See [The Cargo Book:
-//! > External tools] for more on this topic.
+//! <div class="warning">
+//!
+//! This library is maintained by the Cargo team, primarily for use by Cargo
+//! and not intended for external use.
+//! This crate may make major changes to its APIs or assumptions about how it is invoked (e.g.
+//! when it is safe to call [git2::transport::register]).
+//!
+//! See [The Cargo Book: External tools] for alternatives to using this library.
+//!
+//! </div>
 //!
 //! ## Overview
 //!
